requirements.txt
# Broker
confluent-kafka==2.6.1

# Stream processing
duckdb==1.2.2

# Agent
langgraph==0.3.5
langchain-core==0.3.51

# LLM
groq==0.15.0

# API
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1

# HTTP client
httpx==0.28.1

# Utilities
python-dotenv==1.1.0
